Coherent mapping from inductance sensing to driver steps

+++ Software/BLDC/HAL/InductanceSensing.hpp
@@ -6,7 +6,7 @@ namespace HAL {
 namespace BLDC {
 namespace InductanceSensing {
 
-uint16_t RotorPosition();
+int8_t RotorPosition();
 uint32_t WindingInductance();
 
 }
